Refactor: Moved Versionable interface to the other package.
