+++ src/main/java/by/innowise/internship/userService/api/dto/cardInfo/CardInfoResponseDto.java
@@ -1,6 +1,6 @@
 package by.innowise.internship.userService.api.dto.cardInfo;
 
-import by.innowise.internship.userService.core.util.api.Versionable;
+import by.innowise.internship.userService.core.platform.api.Versionable;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.time.LocalDate;
