+++ src/main/java/by/innowise/internship/userService/core/platform/api/Versionable.java
@@ -1,0 +1,6 @@
+package by.innowise.internship.userService.core.platform.api;
+
+public interface Versionable {
+
+    Long getVersion();
+}
