--- src/main/java/by/innowise/internship/userService/core/util/api/Versionable.java
@@ -1,6 +1,0 @@
-package by.innowise.internship.userService.core.util.api;
-
-public interface Versionable {
-
-    Long getVersion();
-}
